v2.0.6
- New version
- Shifted viu install to post
