+++ .npmignore
@@ -6,3 +6,9 @@ api_reference.md
 contribution.md
 _config.yml
 captcha.jpg
+captcha.jpg
+viu
+bin
+viu.exe
+.log
+logs
